requirements.txt
python-libnmap
argparse
termcolor
selenium
requests